Separate sale and sales into separate endpoints.
Finish sale and sales endpoints, tests, and database logic.

Change tax_percent from integer to decimal (numeric).

Add sale validations.

Finish sale PUT edits for editing specific sale items and validating the sale's new total after updating, adding, or removing items.

+++ app/db/tables.js
@@ -21,7 +21,7 @@ module.exports = {
         `sale(
            id SERIAL PRIMARY KEY,
            total integer NOT NULL,
-           tax_percent integer NOT NULL,
+           tax_percent numeric NOT NULL,
            items json NOT NULL,
            modified_date TIMESTAMP,
            created_date TIMESTAMP NOT NULL
